requirements.txt
certifi==2023.7.22
charset-normalizer==3.3.1
idna==3.4
numpy==1.26.2
pandas==2.1.3
Pillow==10.1.0
prompt-toolkit==1.0.14
Pygments==2.16.1
python-dateutil==2.8.2
pytz==2023.3.post1
rapidfuzz==3.5.2
regex==2023.10.3
requests==2.31.0
six==1.16.0
SQLAlchemy==2.0.23
thefuzz==0.20.0
typing_extensions==4.8.0
tzdata==2023.3
urllib3==2.0.7
wcwidth==0.2.10
